% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{representative_cluster}
\alias{representative_cluster}
\title{Title Calculate the representative cluster label from posterior samples}
\usage{
representative_cluster(C, burnin, K)
}
\arguments{
\item{C}{A matrix (N * S) containing posterior samples of labels}

\item{burnin}{The number of samples will be discarded}

\item{K}{The predefined number of clusters}
}
\value{
A vector containing the representative label
}
\description{
Title Calculate the representative cluster label from posterior samples
}
\examples{
representative_cluster(C = matrix(sample(10, 5000, replace = TRUE), nrow = 10), burnin = 250, K = 10)
}
